[BD-8] refactor: domain model 패키지 분리 (dto / enums) 및 Exception 위치 정리 - domain/model/{Artist, Recording, Release, ReleaseGroup, UnifiedSearchHit} → domain/model/dto/ - domain/model/SearchMode → domain/model/enums/SearchMode - adapter/outbound/external/musicbrainz/MusicBrainzApiException → global/error/exception/MusicBrainzApiException - 위 이동에 따른 import 경로 갱신 (controller / mapper / spotify client / port / musicbrainz client)
