+++ src/main/kotlin/com/bandage/musicmetadataservice/application/port/outbound/MusicInfoApiClient.kt
@@ -1,11 +1,11 @@
 package com.bandage.musicmetadataservice.application.port.outbound
 
-import com.bandage.musicmetadataservice.domain.model.Artist
-import com.bandage.musicmetadataservice.domain.model.Recording
-import com.bandage.musicmetadataservice.domain.model.ReleaseGroup
-import com.bandage.musicmetadataservice.domain.model.SearchMode
-import com.bandage.musicmetadataservice.domain.model.SearchSort
-import com.bandage.musicmetadataservice.domain.model.UnifiedSearchHit
+import com.bandage.musicmetadataservice.domain.model.dto.Artist
+import com.bandage.musicmetadataservice.domain.model.dto.Recording
+import com.bandage.musicmetadataservice.domain.model.dto.ReleaseGroup
+import com.bandage.musicmetadataservice.domain.model.enums.SearchMode
+import com.bandage.musicmetadataservice.domain.model.enums.SearchSort
+import com.bandage.musicmetadataservice.domain.model.dto.UnifiedSearchHit
 
 /**
  * 외부 음악 메타데이터 API 호출 포트 (driven port).
